build: move rider-related code to the rider module
