+++ rider/src/main/kotlin/com/jetbrains/aspire/rider/sessions/wasmHost/WasmHostProjectSessionProcessLauncher.kt
@@ -10,7 +10,7 @@ import com.intellij.openapi.project.Project
 import com.jetbrains.rd.util.lifetime.Lifetime
 import com.jetbrains.aspire.run.host.AspireHostConfiguration
 import com.jetbrains.aspire.rider.sessions.DotNetProjectSessionExecutableFactory
-import com.jetbrains.aspire.sessions.findRunnableProjectByPath
+import com.jetbrains.aspire.rider.sessions.findRunnableProjectByPath
 import com.jetbrains.aspire.rider.sessions.projectLaunchers.DotNetSessionWithHotReloadProcessLauncher
 import com.jetbrains.aspire.sessions.DotNetSessionLaunchConfiguration
 import com.jetbrains.rider.nuget.PackageVersionResolution
